added the vectordb
added the vectordb, cuz i think this is what is missing. The code is supposed to create a vectordb, but since it can't i have pre-created it and upload it to github, hopefully this works! Just realised that this includes the sqlite3 too (which was one of the stuff that was missing in the last trial) !!